% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhrsbt.R
\name{rhrsbt}
\alias{rhrsbt}
\title{Relocation Heuristic for Relaxed Structural Balance}
\usage{
rhrsbt(A, C, TLIMIT)
}
\arguments{
\item{A}{An \eqn{N x N} signed network matrix.}

\item{C}{The number of clusters (\eqn{1 < C < N}, where \eqn{N} is the number of nodes).}

\item{TLIMIT}{A desired time limit.}
}
\value{
The function returns the following:
\itemize{
\item \code{obj} - the Doreian & Mrvar objective value;
\item \code{P} - \eqn{N}-dimensional vector of cluser assignements; and
\item \code{restarts} - the number of restarts within the time limit.
}
}
\description{
This function runs relocation heuristic for relaxed structural balance on an \eqn{M x M} asymmetric matrix. The main diagonal is ignored.
}
\examples{
# Load the Sampson (1968) monastery network (3rd time point).
data("sampsonT3")

# Run relocation heuristic for relaxed structural balance.
res <- rhrsbt(A = sampsonT3, C = 3, TLIMIT = 1)

# See the results.
res
}
\references{
Brusco, M. J., Doreian, P., & Steinley, D. (2019). Deterministic blockmodeling of signed and two-mode networks: a tutorial with psychological examples. \emph{British Journal of Mathematical and Statistical Psychology}.

Doreian, P., & Mrvar, A. (2009). Partitioning signed social networks. \emph{Social Networks}, 31, 1-11. http://dx.doi.org/10.1016/j.socnet.2008.08.001
}
\author{
Michael Brusco
}
