% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmklmed.R
\name{tmklmed}
\alias{tmklmed}
\title{Two-Mode Blockmodeling (Structural Equivalence) Heuristic}
\usage{
tmklmed(A, RC, CC, TLIMIT)
}
\arguments{
\item{A}{An \eqn{RO x CO} two-mode network matrix.}

\item{RC}{The number of clusters for row objects (\eqn{1 < RC < RO}).}

\item{CC}{The number of clusters for column objects (\eqn{1 < CC < CO}).}

\item{TLIMIT}{A desired time limit.}
}
\value{
The function returns the following:
\itemize{
\item \code{objval} - total number of inconsistencies;
\item \code{RP} - an \eqn{RO}-dimensional vector of row cluser assignements;
\item \code{RC} - an \eqn{RC}-dimensional vector of column cluser assignements;
\item \code{restarts} - the number of restarts within the time limit.
}
}
\description{
This function runs two-mode KL-medians for an \eqn{RO x CO} network matrix.
}
\examples{
# Load the Galaskiewicz’ (1985) CEO/Club network data.
data("galask")

# Run the two-mode blockmodeling heuristic procedure.
res <- tmklmed(galask, RC = 5, CC = 5, TLIMIT = 1)

# See the results.
res

# Plot the network, in a matrix format, in line with the obtained partitions.
# The function plotMat is from blockmodeling package.
# plotMat(galask, clu = list(res$RP, res$CP))
}
\references{
Brusco, M. J., & Doreian, P. (2019). Partitioning signed networks using relocation heuristics, tabu search, and variable neighborhood search. Social Networks, 56, 70-80. https://doi.org/10.1016/j.socnet.2018.08.007

Doreian, P., & Mrvar, A. (2009). Partitioning signed social networks. Social Networks, 31, 1-11. http://dx.doi.org/10.1016/j.socnet.2008.08.001

Doreian, P., & Mrvar, A. (1996). A partitioning approach to structural balance. Social Networks, 18, 149-168. https://doi.org/10.1016/0378-8733(95)00259-6
}
\author{
Michael Brusco
}
